jarvis/requirements.txt
speech_recognition
pyttsx3
pyautogui
requests
openai
python-dotenv
pyowm
